requirements.txt
streamlit
supabase
gspread
google-auth
python-dotenv
